Clamp the nesting bound in uint64 for 32-bit builds
min(maxDepth, math.MaxUint32) forces the untyped constant into int, which
cannot hold it on a 32-bit platform, so the package failed to compile there.
Comparing in uint64 keeps the clamp exact on every platform; maxDepth is
positive by the time it is clamped, so the conversion is safe.

+++ reflection/common.go
@@ -100,7 +100,7 @@ func NewReflectionCtx(ds sszutils.DynamicSpecs, logCb func(format string, args .
 		noFastSsz:    noFastSsz,
 		noDelegation: noDelegation,
 		maxDepth:     maxDepth,
-		maxLoop:      uint32(min(maxDepth, math.MaxUint32)),
+		maxLoop:      uint32(min(uint64(maxDepth), math.MaxUint32)),
 	}
 }
